Revert "feat(core): widen length path to 32-bit, lifting the 64KB SDU cap"
The 32-bit widening was disproportionate: real UDS reprogramming chunks
payloads via TransferData (each block bounded by the negotiated
maxNumberOfBlockLength and the ECU buffer), so a single >64KB ISO-TP SDU
is not a real use case, and the change forced a breaking API churn across
every length-bearing callback for no practical gain.

Keep the meaningful part (escape FirstFrame + FlowControl WAIT/OVFLW).
The escape format still serves 4096..65535-byte SDUs and answers
FC.OVFLW above the receive buffer, with no API break.

This reverts commit cf2b88c8a026ed26096e31542a68e32790ba6609.

+++ examples/client_demo/main.c
@@ -52,13 +52,13 @@ static uds_isotp_ctx_t g_isotp;
 static uint8_t g_isotp_tx_sdu[1024];
 
 /** Adapter binding the core's fn_tp_send contract to this ISO-TP instance. */
-static int isotp_send_adapter(struct uds_ctx* ctx, const uint8_t* data, uint32_t len)
+static int isotp_send_adapter(struct uds_ctx* ctx, const uint8_t* data, uint16_t len)
 {
     (void) ctx;
     return uds_isotp_send(&g_isotp, data, len);
 }
 
-void on_response(uds_ctx_t* ctx, uint8_t sid, const uint8_t* data, uint32_t len)
+void on_response(uds_ctx_t* ctx, uint8_t sid, const uint8_t* data, uint16_t len)
 {
     (void) ctx;
     printf("[CLIENT] Response Received: SID=%02X, Len=%d\n", sid, len);
