+++ tests/unit/test_service_28.c
@@ -23,7 +23,7 @@ static uint32_t mock_get_time(void)
 {
     return 0;
 }
-static int mock_tp_send(struct uds_ctx *ctx, const uint8_t *data, uint32_t len)
+static int mock_tp_send(struct uds_ctx *ctx, const uint8_t *data, uint16_t len)
 {
     (void) ctx;
     check_expected(data);
